Move lib/yarjuf.rb to lib/j_unit.rb for rspec auto loading without requiring yarjuf gem and keep a simple pointer in lib/yarjuf.rb.
